+++ assembler/src/include/xmath.h
@@ -273,7 +273,7 @@ template<>
 inline double eps<double>() { return 1e-10; }
 
 template<>
-inline float eps<float>() { return 1e-5; }
+inline float eps<float>() { return (float)1e-5; }
 
 template<class T> inline
 bool eq(T lhs, T rhs) {
